# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Post Office Attendance Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List system users
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create system user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemUser"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemUser"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required: [message, success]
...
    SystemUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, regional_admin, branch_admin, viewer]
        branchIds:
          type: array
          items:
            type: integer
        branchNames:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        lastLogin:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, username, fullName, email, role, branchIds, branchNames, isActive, createdAt]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [super_admin, regional_admin, branch_admin, viewer]
        branchIds:
          type: array
          items:
            type: integer
        isActive:
          type: boolean
      required: [username, fullName, email, password, role, branchIds, isActive]
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
          nullable: true
        role:
          type: string
          enum: [super_admin, regional_admin, branch_admin, viewer]
        branchIds:
          type: array
          items:
            type: integer
        isActive:
          type: boolean
      required: [fullName, email, role, branchIds, isActive]
...